+++ samples/AfterContosoUniversity/ContosoUniversity/Home.razor.cs
@@ -1,9 +1,0 @@
-using Microsoft.AspNetCore.Components;
-
-namespace ContosoUniversity.ContosoUniversity
-{
-    public partial class Home : ComponentBase
-    {
-        // Home page is static - no data access required
-    }
-}