+++ samples/AfterContosoUniversity/Home.razor.cs
@@ -1,0 +1,11 @@
+using Microsoft.AspNetCore.Components;
+
+namespace ContosoUniversity;
+
+public partial class Home : ComponentBase
+{
+    protected override Task OnInitializedAsync()
+    {
+        return Task.CompletedTask;
+    }
+}